+++ src/models/incidentals/one_time_incidentals.ts
@@ -39,6 +39,11 @@ interface OneTimeIncidentals {
    * Document id of the bill
    */
   billDocumentId?: string;
+
+  /**
+   * Document id of the bank transfer
+   */
+  bankTransferDocumentId?: string;
 }
 
 export default OneTimeIncidentals;